+++ packages/core/nuxt-theme-module/theme/pages/Category.vue
@@ -11,6 +11,7 @@
       <div class="navbar__main">
         <SfButton
           class="sf-button--text navbar__filters-button"
+          data-cy="category-btn_filters"
           aria-label="Filters"
           @click="toggleFilterSidebar"
         >
@@ -19,6 +20,7 @@
             color="dark-secondary"
             icon="filter2"
             class="navbar__filters-icon"
+            data-cy="category-icon_"
           />
            {{ $t('Filters') }}
         </SfButton>
@@ -59,6 +61,7 @@
             @click="toggleCategoryGridView"
           />
           <SfIcon
+            data-cy="category-icon_list-view"
             class="navbar__view-icon"
             :color="!isCategoryGridView ? 'black' : 'dark-secondary'"
             icon="list"
@@ -236,6 +239,81 @@
         </div>
       </SfLoader>
     </div>
+    <SfSidebar
+      :visible="isFilterSidebarOpen"
+      title="Filters"
+      class="sidebar-filters"
+      @close="toggleFilterSidebar"
+    >
+      <div class="filters desktop-only">
+        <div v-for="(facet, i) in facets" :key="i">
+          <SfHeading
+            :level="4"
+            :title="facet.label"
+            class="filters__title sf-heading--left"
+            :key="`filter-title-${facet.id}`"
+          />
+            <div
+              v-if="isFacetColor(facet)"
+              class="filters__colors"
+              :key="`${facet.id}-colors`"
+            >
+              <SfColor
+                v-for="option in facet.options"
+                :key="`${facet.id}-${option.value}`"
+                :data-cy="`category-filter_color_${option.value}`"
+                :color="option.value"
+                :selected="isFilterSelected(facet, option)"
+                class="filters__color"
+                @click="() => selectFilter(facet, option)"
+              />
+            </div>
+            <div v-else>
+              <SfFilter
+                v-for="option in facet.options"
+                :key="`${facet.id}-${option.value}`"
+                :data-cy="`category-filter_${facet.id}_${option.value}`"
+                :label="option.id + `${option.count && ` (${option.count})`}`"
+                :selected="isFilterSelected(facet, option)"
+                class="filters__item"
+                @change="() => selectFilter(facet, option)"
+              />
+            </div>
+        </div>
+      </div>
+      <SfAccordion class="filters smartphone-only">
+        <div v-for="(facet, i) in facets" :key="i">
+          <SfAccordionItem
+            :key="`filter-title-${facet.id}`"
+            :header="facet.label"
+            class="filters__accordion-item"
+          >
+          <SfFilter
+            v-for="option in facet.options"
+            :key="`${facet.id}-${option.id}`"
+            :label="option.id"
+            :selected="isFilterSelected(facet, option)"
+            class="filters__item"
+            @change="option.selected = !option.selected"
+          />
+        </SfAccordionItem>
+      </div>
+      </SfAccordion>
+      <template #content-bottom>
+        <div class="filters__buttons">
+          <SfButton
+            class="sf-button--full-width"
+            @click="applyFilters"
+            >Done</SfButton
+          >
+          <SfButton
+            class="sf-button--full-width filters__button-clear"
+            @click="clearFilters"
+            >Clear all</SfButton
+          >
+        </div>
+      </template>
+    </SfSidebar>
   </div>
 </template>
 
@@ -258,16 +336,15 @@ import {
   SfColor,
   SfProperty
 } from '@storefront-ui/vue';
-import { computed, onMounted } from '@vue/composition-api';
+import { ref, computed, onMounted } from '@vue/composition-api';
 import { useCart, useFacet, useWishlist, facetGetters, productGetters } from '<%= options.generate.replace.composables %>';
 import { useUiHelpers, useUiState } from '~/composables';
 import { onSSR } from '@vue-storefront/core';
-import Filters from '../components/Filters';
+import Vue from 'vue';
 
 export default {
   transition: 'fade',
   setup(props, context) {
-    onMounted(() => context.root.$scrollTo(context.root.$el, 2000));
     const th = useUiHelpers();
     const uiState = useUiState();
     const { addToCart, isOnCart } = useCart();
@@ -284,6 +361,47 @@ export default {
       await search(th.getFacetsFromURL());
     });
 
+    const { changeFilters, isFacetColor } = useUiHelpers();
+    const { toggleFilterSidebar } = useUiState();
+    const selectedFilters = ref({});
+
+    onMounted(() => {
+      context.root.$scrollTo(context.root.$el, 2000);
+      if (!facets.value.length) return;
+      selectedFilters.value = facets.reduce((prev, curr) => ({
+        ...prev,
+        [curr.id]: curr.options
+          .filter(o => o.selected)
+          .map(o => o.id)
+      }), {});
+    });
+
+    const isFilterSelected = (facet, option) => (selectedFilters.value[facet.id] || []).includes(option.id);
+
+    const selectFilter = (facet, option) => {
+      if (!selectedFilters.value[facet.id]) {
+        Vue.set(selectedFilters.value, facet.id, []);
+      }
+
+      if (selectedFilters.value[facet.id].find(f => f === option.id)) {
+        selectedFilters.value[facet.id] = selectedFilters.value[facet.id].filter(f => f !== option.id);
+        return;
+      }
+
+      selectedFilters.value[facet.id].push(option.id);
+    };
+
+    const clearFilters = () => {
+      toggleFilterSidebar();
+      selectedFilters.value = {};
+      changeFilters(selectedFilters.value);
+    };
+
+    const applyFilters = () => {
+      toggleFilterSidebar();
+      changeFilters(selectedFilters.value);
+    };
+
     return {
       ...uiState,
       th,
@@ -297,7 +415,13 @@ export default {
       breadcrumbs,
       addToWishlist,
       addToCart,
-      isOnCart
+      isOnCart,
+      isFacetColor,
+      selectFilter,
+      isFilterSelected,
+      selectedFilters,
+      clearFilters,
+      applyFilters
     };
   },
   components: {
@@ -316,7 +440,6 @@ export default {
     SfLoader,
     SfColor,
     SfHeading,
-    Filters,
     SfProperty
   }
 };
@@ -419,13 +542,12 @@ export default {
     --select-error-message-height: 0;
     ::v-deep .sf-select__dropdown {
       font-size: var(--font-size-sm);
-      font-family: var(--font-family--primary);
+      font-family: var(--font-family--secondary);
       font-weight: var(--font-weight--light);
       margin: 0;
     }
     ::v-deep .sf-select__placeholder {
       --select-option-font-size: var(--font-size-sm);
-      color: red;
     }
   }
   &__sort {
@@ -605,248 +727,4 @@ export default {
     margin: var(--spacer-xs) 0 0 0;
   }
 }
-// #category {
-//   box-sizing: border-box;
-//   @include for-desktop {
-//     max-width: 1240px;
-//     margin: 0 auto;
-//   }
-// }
-// .main {
-//   &.section {
-//     padding: var(--spacer-xs);
-//     @include for-desktop {
-//       padding: 0;
-//     }
-//   }
-// }
-// .breadcrumbs {
-//   padding: var(--spacer-base) 0;
-// }
-// .navbar {
-//   position: relative;
-//   display: flex;
-//   border: 1px solid var(--c-light);
-//   border-width: 0 0 1px 0;
-//   @include for-desktop {
-//     border-width: 1px 0 1px 0;
-//   }
-//   &.section {
-//     padding: var(--spacer-sm);
-//     @include for-desktop {
-//       padding: 0;
-//     }
-//   }
-//   &__aside,
-//   &__main {
-//     display: flex;
-//     align-items: center;
-//     padding: var(--spacer-sm) 0;
-//     flex: 1;
-//     padding: 0;
-//     @include for-desktop {
-//       padding: var(--spacer-xs) var(--spacer-xl);
-//     }
-//   }
-//   &__aside {
-//     flex: 0 0 15%;
-//     padding: var(--spacer-sm) var(--spacer-sm);
-//     border: 1px solid var(--c-light);
-//     border-width: 0 1px 0 0;
-//   }
-//   &__title {
-//     --heading-title-font-weight: var(--font-weight--light);
-//     --heading-title-font-size: var(--font-size--xl);
-//   }
-//   &__filters-icon {
-//     margin: 0 var(--spacer-sm) 0 0;
-//   }
-//   &__filters-button {
-//     --button-color: var(--c-link);
-//     --button-font-size: var(--font-size--base);
-//     --button-font-weight: var(--font-weight--normal);
-//     --button-text-decoration: none;
-//     display: flex;
-//     align-items: center;
-//     @include for-mobile {
-//       order: 1;
-//     }
-//     svg {
-//       fill: var(--c-text-muted);
-//       transition: fill 150ms ease;
-//     }
-//     &:hover {
-//       svg {
-//         fill: var(--c-primary);
-//       }
-//     }
-//   }
-//   &__label {
-//     font-family: var(--font-family--secondary);
-//     font-weight: var(--font-weight--normal);
-//     color: var(--c-link);
-//     margin: 0 var(--spacer-xs) 0 0;
-//   }
-//   &__select {
-//     --select-option-font-size: var(--font-size--base);
-//     --select-padding: var(--spacer-sm) 0;
-//     ::v-deep .sf-select__dropdown {
-//       font-size: var(--font-size--base);
-//       margin: var(--spacer-2xs) 0 0 0;
-//     }
-//   }
-//   &__sort {
-//     display: flex;
-//     align-items: center;
-//     margin: 0 auto 0 var(--spacer-2xl);
-//   }
-//   &__counter {
-//     font-family: var(--font-family--secondary);
-//     margin: auto;
-//     @include for-desktop {
-//       margin: auto 0 auto auto;
-//     }
-//   }
-//   &__view {
-//     display: flex;
-//     align-items: center;
-//     @include for-desktop {
-//       margin: 0 0 0 var(--spacer-2xl);
-//     }
-//     @include for-mobile {
-//       order: -1;
-//     }
-//     &-icon {
-//       cursor: pointer;
-//       margin: 0 var(--spacer-base) 0 0;
-//       &:last-child {
-//         margin: 0;
-//       }
-//     }
-//     &-label {
-//       margin: 0 var(--spacer-sm) 0 0;
-//       font: var(--font-weight--normal) var(--font-size--base) / 1.6 var(--font-family--secondary);
-//       color: var(--c-link);
-//     }
-//   }
-// }
-// .sort-by {
-//   --select-dropdown-z-index: 1;
-//   flex: unset;
-//   width: 11.875rem;
-// }
-// .main {
-//   display: flex;
-//   .loading {
-//     margin-top: 100px;
-//   }
-// }
-// .sidebar {
-//   flex: 0 0 15%;
-//   padding: var(--spacer-sm);
-//   border: 1px solid var(--c-light);
-//   border-width: 0 1px 0 0;
-//   .loading {
-//     margin-top: 50px;
-//   }
-// }
-// .sidebar-filters {
-//   --sidebar-title-display: none;
-//   --sidebar-top-padding: 0;
-//   @include for-desktop {
-//     --sidebar-content-padding: 0 var(--spacer-xl);
-//     --sidebar-bottom-padding: 0 var(--spacer-xl);
-//   }
-// }
-// .list {
-//   --menu-item-font-size: var(--font-sm);
-//   &__item {
-//     &:not(:last-of-type) {
-//       --list-item-margin: 0 0 var(--spacer-sm) 0;
-//     }
-//   }
-// }
-// .products {
-//   box-sizing: border-box;
-//   flex: 1;
-//   margin: 0;
-//   &__grid,
-//   &__list {
-//     display: flex;
-//     flex-wrap: wrap;
-//   }
-//   &__grid {
-//     justify-content: space-between;
-//   }
-//   &__grid,
-//   &__list {
-//     display: flex;
-//     flex-wrap: wrap;
-//     justify-content: center;
-//     @include for-desktop {
-//       justify-content: flex-start;
-//     }
-//   }
-//   &__product-card {
-//     --product-card-add-button-transform: translate3d(0, 30%, 0);
-//      flex: 1 1 50%;
-//     @include for-desktop {
-//       --product-card-padding: var(--spacer-sm);
-//       flex: 1 1 25%;
-//     }
-//   }
-//   &__product-card-horizontal {
-//     flex: 0 0 100%;
-//   }
-//   &__slide-enter {
-//     opacity: 0;
-//     transform: scale(0.5);
-//   }
-//   &__slide-enter-active {
-//     transition: all 0.2s ease;
-//     transition-delay: calc(0.1s * var(--index));
-//   }
-//   &__pagination {
-//       display: flex;
-//       justify-content: center;
-//       margin: var(--spacer-xl) 0 0 0;
-//   }
-//   @include for-desktop {
-//     margin: var(--spacer-sm) 0 0 var(--spacer-sm);
-//     &__pagination {
-//       &__options {
-//         display: flex;
-//         align-items: baseline;
-//         justify-content: flex-end;
-//       }
-//       &__label {
-//         font-family: var(--font-family--secondary);
-//         font-size: var(--font-size--sm);
-//         font-weight: var(--font-weight--normal);
-//         color: var(--c-link);
-//         margin-right: var(--spacer-2xs);
-//       }
-//     }
-//     &__product-card-horizontal {
-//       margin: var(--spacer-lg) 0;
-//     }
-//     &__product-card {
-//       flex: 1 1 25%;
-//     }
-//     &__list {
-//       margin: 0 0 0 var(--spacer-sm);
-//     }
-//   }
-// }
-// .filters {
-//   &__accordion-item {
-//     --accordion-item-content-padding: 0;
-//     position: relative;
-//     left: 50%;
-//     right: 50%;
-//     margin-left: -50vw;
-//     margin-right: -50vw;
-//     width: 100vw;
-//   }
-// }
 </style>
